upgrade contentful deps
