+++ .nvmrc
@@ -1,1 +1,1 @@
-20.12.2
+20.18.1
