+++ package.json
@@ -6,24 +6,23 @@
   "type": "module",
   "license": "MIT",
   "engines": {
-    "node": ">=20.12.2"
+    "node": ">=20.18.1"
   },
   "scripts": {
+    "clean": "rimraf ./public .greenwood",
     "prisma": "prisma",
     "lint": "eslint \"*.*js\" \"./src/**/**/*.*js\"",
     "dev": "node --env-file=.env ./node_modules/@greenwood/cli/src/index.js develop",
     "build": "node --env-file=.env ./node_modules/@greenwood/cli/src/index.js build",
-    "serve": "npm run clean && npm run build && greenwood serve",
+    "serve": "npm run clean && npm run build && node --env-file=.env ./node_modules/@greenwood/cli/src/index.js serve",
     "studio": "prisma studio"
   },
-  "dependencies": {
-    "@aws-sdk/client-cloudfront": "^3.58.0",
-    "@contentful/rich-text-html-renderer": "^15.11.1",
-    "@libsql/client": "^0.6.0",
-    "contentful": "^9.1.18"
-  },
   "devDependencies": {
+    "@aws-sdk/client-cloudfront": "^3.58.0",
+    "@contentful/rich-text-html-renderer": "^17.0.0",
     "@greenwood/cli": "^0.31.0-alpha.0",
+    "@libsql/client": "^0.6.0",
+    "contentful": "11.3.0",
     "eslint": "^8.11.0",
     "prisma": "^5.7.0",
     "rimraf": "^5.0.5"
