+++ src/pages/api/albums.js
@@ -1,15 +1,18 @@
-import mysql from 'mysql2/promise';
+import { createClient } from '@libsql/client/web';
 
 export async function handler (request) {
   const params = new URLSearchParams(request.url.slice(request.url.indexOf('?')));
   const id = params.has('id') ? params.get('id') : null;
   const artistId = params.has('artistId') ? params.get('artistId') : null;
-  const connection = await mysql.createConnection(process.env.DATABASE_URL);
-  const [rows] = id
-    ? await connection.execute('SELECT * FROM albums WHERE id = ?', [id])
+  const client = createClient({
+    url: process.env.DATABASE_URL,
+    authToken: process.env.DATABASE_TOKEN
+  });
+  const { rows } = id
+    ? await client.execute({ sql: 'SELECT * FROM albums WHERE id = ?', args: [id] })
     : artistId
-      ? await connection.execute('SELECT * FROM albums WHERE artistId = ?', [artistId])
-      : await connection.execute('SELECT * FROM albums');
+      ? await client.execute({ sql: 'SELECT * FROM albums WHERE artistId = ?', args: [artistId] })
+      : await client.execute('SELECT * FROM albums');
 
   return new Response(JSON.stringify(rows), {
     headers: new Headers({
