+++ src/pages/api/posts.js
@@ -1,12 +1,15 @@
-import mysql from 'mysql2/promise';
+import { createClient } from '@libsql/client/web';
 
 export async function handler (request) {
   const params = new URLSearchParams(request.url.slice(request.url.indexOf('?')));
   const id = params.has('id') ? params.get('id') : null;
-  const connection = await mysql.createConnection(process.env.DATABASE_URL);
-  const [rows] = id
-    ? await connection.execute('SELECT * FROM posts WHERE id = ?', [id])
-    : await connection.execute('SELECT * FROM posts');
+  const client = createClient({
+    url: process.env.DATABASE_URL,
+    authToken: process.env.DATABASE_TOKEN
+  });
+  const { rows } = id
+    ? await client.execute({ sql: 'SELECT * FROM posts WHERE id = ?', args: [id] })
+    : await client.execute('SELECT * FROM posts');
 
   return new Response(JSON.stringify(rows), {
     headers: new Headers({
