+++ src/pages/index.html
@@ -60,6 +60,7 @@
         width: 10%;
         height: 50px;
         border-radius: 5px;
+        cursor: pointer;
       }
       details {
         margin: 20px auto;
@@ -74,7 +75,7 @@
   </head>
 
   <body>
-    <h1>Resource Logger</h1>
+    <h1>Endpoint Debugger</h1>
     <h2>Number of results: <span id="results"></span></h2>
     <button id="albums">Albums</button>
     <button id="artists">Artists</button>
